Trunghau

+++ src/components/teacher/details/common/answerQuestion/useQuestion.js
@@ -5,7 +5,7 @@ import {
   deleteAnswer,
   updateAnswer,
   updateQuestion,
-} from "../../../../../api/teacher/answerQuestionService";
+} from "api/study/answerQuestion/answerQuestionService";
 
 export default function useQuestion(data, fetchData, setError, path) {
   const [question, setQuestion] = useState(data);
@@ -98,8 +98,15 @@ export default function useQuestion(data, fetchData, setError, path) {
     setQuestion({ ...question, answers: newAnswers });
   }
 
-  async function onDeleteAnswer(id) {
+  async function onDeleteAnswer(id, index) {
     if (!isEditing) return;
+    if (id === "-1") {
+      setQuestion((prevData) => ({
+        ...prevData,
+        answers: prevData.answers.filter((_, i) => i !== index),
+      }));
+      return;
+    }
     try {
       await deleteAnswer(path, id);
       await fetchData();
